Chat emote update 1.1.0, and update to 1.19.

+++ src/main/java/tensor/mixin/feature/ItemUseCooldown_MinecraftClient.java
@@ -1,15 +1,11 @@
 package tensor.mixin.feature;
 
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.text.TranslatableText;
 import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import tensor.Tensor;
 import tensor.option.TensorOptions;
 
 @Mixin(MinecraftClient.class)
@@ -22,7 +18,7 @@ public class ItemUseCooldown_MinecraftClient
     private void changeItemUse(MinecraftClient client, int cooldown)
     {
         if(TensorOptions.itemUseCooldownOverrideToggle)
-            itemUseCooldown = (int) TensorOptions.itemUseCooldown;
+            itemUseCooldown = TensorOptions.itemUseCooldown;
         else
             itemUseCooldown = 4;
     }
